fix lint issues

+++ app/src/main/java/com/example/medregister/RegisterPillsActivity.java
@@ -141,13 +141,11 @@ public class RegisterPillsActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.delete_all_pills:
-                pillViewModel.deleteAllPills();
-                Toast.makeText(this, R.string.all_pills_deleted, Toast.LENGTH_SHORT).show();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        if (item.getItemId() == R.id.delete_all_pills) {
+            pillViewModel.deleteAllPills();
+            Toast.makeText(this, R.string.all_pills_deleted, Toast.LENGTH_SHORT).show();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 }
